+++ src/features/temptation/pages/TemptationMain.module.css
